backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
motor==3.6.0
pymongo==4.9.1
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
argon2-cffi==23.1.0
python-multipart==0.0.6
python-dotenv==1.0.0
email-validator==2.3.0
certifi==2025.11.12
